+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to `bankly-laravel` will be documented in this file
 
+## 1.18.1 - 2021-10-19
+- Fixed bankly pix facade class name
+
+## 1.18.0 - 2021-09-29
+- Create static QRCodes
+- Reading static and dynamic QRCodes
+
+## 1.17.0 - 2021-09-23
+- Create customer account
+
 ## 1.16.0 - 2021-08-26
 - Allows sending PIX to payment type accounts